Authorized deletion of a post using the token created!

+++ controllers/api/v1/posts_api.js
@@ -30,8 +30,7 @@ module.exports.destroy = async function(req, res) {
     try {
         let post = await Post.findById(req.params.id);
 
-        // .id means converting the object id into string
-        // if(post.user == req.user.id) {
+        if(post.user == req.user.id) {
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
@@ -39,11 +38,12 @@ module.exports.destroy = async function(req, res) {
             return res.status(200).json({
                 message: "Post and associated comments deleted successfully!"
             });
-        // }
-        // else {
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+        }
+        else {
+            return res.status(401).json({
+                message: "Unauthorized to delete this post!"
+            });
+        }
     }
     catch(err) {
         console.log('******', err);
